requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
coloredlogs==15.0.1
distro==1.9.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
humanfriendly==10.0
idna==3.6
install==1.3.5
openai==1.12.0
pillow==10.2.0
pydantic==2.6.2
pydantic_core==2.16.3
python-dotenv==1.0.1
requests==2.31.0
slack-bolt==1.18.1
slack_sdk==3.27.0
sniffio==1.3.0
tqdm==4.66.2
typing_extensions==4.9.0
Unidecode==1.3.8
urllib3==2.2.1
